requirements.txt
langchain==0.3.4
jupyter==1.1.1
langchain-groq==0.2.0
python-dotenv==1.0.1
langserve[all]
langchain-community==0.3.3
langchain-core==0.3.12
langgraph>0.2.27
unstructured==0.16.2
langchain_unstructured==0.1.5
unstructured[pdf]
langchain-huggingface==0.1.0
pydantic==2.9.2
pinecone-client[grpc]
langchain-pinecone==0.2.0
beautifulsoup4==4.12.3
requests==2.32.3
einops==0.8.0
tqdm
fastapi[standard]==0.115.3
psycopg==3.2.3
psycopg2==2.9.10
python-jose[cryptography]==3.3.0
python-multipart==0.0.16
authlib==1.3.2
itsdangerous==2.2.0
starlette==0.41.2
starlette-session==0.4.3
psycopg[pool]
langgraph-checkpoint-postgres==2.0.2
sqlmodel==0.0.22
firebase-admin==6.6.0
